+++ yggdrasil_decision_forests/port/python/CHANGELOG.md
@@ -15,6 +15,7 @@
 -   Warn if training on a VerticalDataset and fail if attempting to modify the
     columns in a VerticalDataset during training.
 -   User can override the model's task, label or group during evaluation.
+-   Add `num_examples_per_tree()` method to Isolation Forest models.
 
 ## 0.7.0 - 2024-08-21